services/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
psycopg2-binary==2.9.9
httpx==0.25.1
websockets==12.0
python-dotenv==1.0.0
alembic==1.12.1 